requirements.txt
boto3==1.38.18
s3transfer==0.12.0
botocore==1.38.18
jmespath==1.0.1
Pillow==9.5.0
moviepy==1.0.3
decorator==4.4.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
proglog==0.1.12
protobuf==6.30.2
google-generativeai==1.14.0
httpx==0.28.1
duckduckgo-search==8.0.1
lxml==5.4.0
primp==0.15.0
bing-image-downloader==1.1.2
google-images-download==2.8.0
google-images-downloader==1.0.16
selenium==4.31.0
trio-websocket==0.12.2
trio==0.30.0
outcome==1.3.0.post0
PySocks==1.7.1
wsproto==1.2.0
sortedcontainers==2.4.0
openai-whisper
moviepy~=1.0.3
pydub~=0.25.1
python-dotenv~=1.1.0
Pillow~=9.5.0